refactor: split relay.go at the seams it already had
internal/relay/relay.go had grown to 992 lines against the project's 800
ceiling, and the two halves that made it long were already contiguous blocks
at the end of the file:

- roster.go: which sessions a caller may address and how a typed name becomes
  one of them — candidate, roster, resolve, matching, and the sentences an
  unresolvable name gets back.
- compose.go: the message typed at the target's prompt and the keystrokes that
  put it there — compose, replyInstruction, unreadNotice, origin, paste,
  submit, the submit delay and sanitize.

The core keeps the Service, the ticket and everything that moves one: Send,
Wait, Reply, Observe, the receipt watch and the marks. Every function moved
whole, with the comments that carry its reasons; nothing was reworded, and
the only line that is not a move is the strconv import, which followed
knownLabels out. 743 lines left here, 155 and 118 in the new files.
