+++ internal/relay/relay.go
@@ -18,7 +18,6 @@ import (
 	"encoding/hex"
 	"fmt"
 	"log/slog"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -728,255 +727,6 @@ func (s *Service) sweep() []*ticket {
 		}
 	}
 	return expired
-}
-
-// candidate is a resolved peer together with the session id behind it, which
-// callers never see: they address a session by the label on its card.
-type candidate struct {
-	ID   string
-	Peer Peer
-}
-
-// roster returns every live session except the caller's own.
-func (s *Service) roster(fromID string) ([]candidate, error) {
-	projects, err := s.sessions.LoadState()
-	if err != nil {
-		return nil, fmt.Errorf("read sessions: %w", err)
-	}
-	var found []candidate
-	for _, p := range projects {
-		for _, sess := range p.Sessions {
-			if sess.ID == fromID || !s.term.Live(sess.ID) {
-				continue
-			}
-			cwd := sess.Path
-			if cwd == "" {
-				cwd = p.Path
-			}
-			found = append(found, candidate{
-				ID: sess.ID,
-				Peer: Peer{
-					Label:   sess.Label,
-					Name:    RosterName(cwd, sess.ID),
-					Project: p.Name,
-					Kind:    sess.Kind,
-				},
-			})
-		}
-	}
-	return found, nil
-}
-
-// resolve finds the single live session named by target, optionally narrowed to
-// one project. An ambiguous label is an error naming every match, because
-// guessing which session a prompt lands in is the one mistake this feature must
-// not make.
-func (s *Service) resolve(fromID, target, project string) (candidate, error) {
-	target = strings.TrimSpace(target)
-	if target == "" {
-		return candidate{}, fmt.Errorf("no target session given")
-	}
-	found, err := s.roster(fromID)
-	if err != nil {
-		return candidate{}, err
-	}
-
-	// The label first, then the roster name. Both address this session, and an
-	// agent may have been handed either — a mention writes the roster name at
-	// the prompt, list_sessions prints both. The label wins a tie because it is
-	// the name the user chose and the one every message and error quotes.
-	matches := matching(found, target, project, func(c candidate) string { return c.Peer.Label })
-	if len(matches) == 0 {
-		matches = matching(found, target, project, func(c candidate) string { return c.Peer.Name })
-	}
-
-	switch len(matches) {
-	case 1:
-		return matches[0], nil
-	case 0:
-		return candidate{}, fmt.Errorf(
-			"no live session named %q%s. %s",
-			target, inProject(project), knownLabels(found),
-		)
-	default:
-		return candidate{}, fmt.Errorf(
-			"%q names %d live sessions (%s) — narrow it with the project",
-			target, len(matches), projectsOf(matches),
-		)
-	}
-}
-
-// labelOf names the sending session for the message it is about to deliver.
-// An empty id is a caller with no session at all — the CLI run from a plain
-// shell or a script — and stays empty, which compose words differently. A
-// caller lich has no record of still gets to send: the receiving agent is told
-// the sender is unknown rather than told nothing at all.
-func (s *Service) labelOf(id string) string {
-	if id == "" {
-		return ""
-	}
-	projects, err := s.sessions.LoadState()
-	if err != nil {
-		return "unknown"
-	}
-	for _, p := range projects {
-		for _, sess := range p.Sessions {
-			if sess.ID == id {
-				return sess.Label
-			}
-		}
-	}
-	return "unknown"
-}
-
-// matching returns the candidates whose name — read out by naming — is target,
-// narrowed to one project when one was given.
-func matching(found []candidate, target, project string, naming func(candidate) string) []candidate {
-	var matches []candidate
-	for _, c := range found {
-		if !strings.EqualFold(naming(c), target) {
-			continue
-		}
-		if project != "" && !strings.EqualFold(c.Peer.Project, project) {
-			continue
-		}
-		matches = append(matches, c)
-	}
-	return matches
-}
-
-// knownLabels names what is actually reachable, under both names, so a wrong
-// name is one step from a right one rather than a guess.
-func knownLabels(found []candidate) string {
-	if len(found) == 0 {
-		return "No other session is live."
-	}
-	names := make([]string, 0, len(found))
-	for _, c := range found {
-		names = append(names, fmt.Sprintf("%s (%s)", strconv.Quote(c.Peer.Label), c.Peer.Name))
-	}
-	return "Live sessions: " + strings.Join(names, ", ") + "."
-}
-
-func inProject(project string) string {
-	if project == "" {
-		return ""
-	}
-	return fmt.Sprintf(" in project %q", project)
-}
-
-func projectsOf(matches []candidate) string {
-	names := make([]string, 0, len(matches))
-	for _, c := range matches {
-		names = append(names, c.Peer.Project)
-	}
-	return strings.Join(names, ", ")
-}
-
-// compose is the message typed at the target's prompt. It names where the
-// request came from so the receiving agent knows this did not come from the
-// person in front of it, and carries the exact command that sends an answer
-// home — the reply path only exists because this text describes it, so the
-// agent needs no prior knowledge of the feature.
-func compose(sender, ticketID, prompt string, hasTools bool) string {
-	return fmt.Sprintf(
-		"[lich] %s, not from your own prompt.\n\n%s\n\n%s",
-		origin(sender), prompt, replyInstruction(hasTools, ticketID),
-	)
-}
-
-// replyInstruction tells the receiving agent how to answer. Every agent has a
-// shell, so the command is always named; a provider lich registers its MCP
-// server with is offered the tool first, because a session that withholds shell
-// access would otherwise have no way to answer at all.
-//
-// The exclusivity is load-bearing, not emphasis. A Claude Code session has a
-// peer channel of its own and this message names another session, so answering
-// through that channel is the obvious move — and it reaches a sender that is
-// blocked on a ticket instead. That happened on the first real run: the target
-// replied over its own socket and the errand timed out with the answer already
-// written. The ticket is the only route home, and the message has to say so.
-func replyInstruction(hasTools bool, ticketID string) string {
-	command := fmt.Sprintf("  \"$LICH_BIN\" reply %s \"<your answer>\"", ticketID)
-	route := "When you have an answer, send it back by running:\n" + command
-	if hasTools {
-		route = fmt.Sprintf(
-			"When you have an answer, send it back with the lich tool `%s` (ticket %s), or by running:\n%s",
-			ToolReply, ticketID, command,
-		)
-	}
-	return route + "\n\nThat ticket is the only way back: whoever asked is blocked on it and " +
-		"is reading nothing else. Do not answer by messaging a peer session — an answer " +
-		"sent any other way is lost."
-}
-
-// unreadNotice is what a sender who stopped waiting is told, typed at its own
-// prompt the way an answer would be. It has to be actionable on its own: the
-// sender cannot see the other screen, and what is usually on it is a question
-// only a person can answer.
-func unreadNotice(target string) string {
-	return fmt.Sprintf(
-		"[lich] The %q session never picked up the task you sent it. It was typed at "+
-			"that prompt and nothing read it, so something else has that terminal — a "+
-			"provider still starting, or a question of its own on screen. Nothing is "+
-			"queued and nothing was answered.",
-		target,
-	)
-}
-
-// origin describes the sender in the message's first line. An empty sender is
-// the lich CLI run outside any session — a script, a scheduled job, the user's
-// own shell — which is a different thing to be told than "another agent".
-func origin(sender string) string {
-	if sender == "" {
-		return "Message relayed by the lich command line"
-	}
-	return fmt.Sprintf("Message from session %q", sender)
-}
-
-// paste wraps text in bracketed paste, which is how a multi-line message
-// reaches a TUI prompt as one prompt instead of as one submission per newline.
-// Every provider lich spawns runs a TUI that enables bracketed paste; one that
-// did not would read the newlines as submissions.
-//
-// It does not submit. See submitDelay.
-func paste(text string) string {
-	return "\x1b[200~" + text + "\x1b[201~"
-}
-
-// submit is the Enter that sends what paste put at the prompt.
-const submit = "\r"
-
-// defaultSubmitDelay is how long the relay waits between the paste and the
-// Enter that sends it.
-//
-// Everything else lich pastes into a prompt is left for the user to send, so
-// this is the only place that presses Enter itself — and a carriage return
-// riding in the same write as the paste is swallowed. Claude Code collapses a
-// multi-line paste into a "[Pasted text #2 +7 lines]" placeholder, and the
-// Enter arriving inside that same burst goes into building the placeholder
-// rather than sending it: the message sat unsent at the target's prompt, seen
-// only when someone opened that session by hand. Nothing here can read the
-// target's screen to know when it has settled, so the delay is the instrument,
-// and it is generous on purpose — a tenth of a second nobody notices against a
-// message that otherwise never arrives.
-const defaultSubmitDelay = 150 * time.Millisecond
-
-// sanitize strips the control characters that would either break out of the
-// bracketed paste framing or drive the target's terminal, keeping the
-// whitespace a prompt legitimately contains. ESC is the one that matters: text
-// carrying "\x1b[201~" would end the paste early and leave the rest of itself
-// running as keystrokes.
-func sanitize(text string) string {
-	return strings.Map(func(r rune) rune {
-		if r == '\n' || r == '\t' {
-			return r
-		}
-		if r < 0x20 || r == 0x7f {
-			return -1
-		}
-		return r
-	}, text)
 }
 
 // ticketIDBytes is the length of a ticket id before hex encoding. Short enough
